added eclipse project
