+++ .gitignore
@@ -15,3 +15,4 @@ bundle.js
 constants.json
 parser/target
 !test/*
+/bin/
